refactor the inheritance
