+++ mmengine/_strategy/single_device.py
@@ -1,11 +1,14 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import Dict, List, Optional, Union
+import time
+from typing import Callable, Dict, List, Optional, Union
 
 import torch.nn as nn
 
+import mmengine
 from mmengine.device import get_device
 from mmengine.optim import OptimWrapper, _ParamScheduler
 from mmengine.registry import STRATEGIES
+from mmengine.utils import get_git_hash
 from .base import BaseStrategy
 
 
@@ -76,3 +79,162 @@ class SingleDeviceStrategy(BaseStrategy):
         model = self.convert_model(model)
         current_device = get_device()
         return model.to(current_device)
+
+    def load_checkpoint(
+        self,
+        filename: str,
+        *,
+        map_location: Union[str, Callable] = 'cpu',
+        callback: Optional[Callable] = None,
+    ) -> dict:
+        """Load checkpoint from given ``filename``.
+
+        Args:
+            filename (str): Accept local filepath, URL, ``torchvision://xxx``,
+                ``open-mmlab://xxx``.
+            map_location (str or callable): A string or a callable function to
+                specifying how to remap storage locations.
+                Defaults to 'cpu'.
+        """
+        from mmengine.runner.checkpoint import _load_checkpoint
+
+        if hasattr(self, 'extra_ckpt'):
+            return self.extra_ckpt
+
+        self.logger.info(f'Load checkpoint from {filename}')
+
+        if map_location == 'default':
+            device = get_device()
+            self.extra_ckpt = _load_checkpoint(filename, map_location=device)
+        else:
+            self.extra_ckpt = _load_checkpoint(
+                filename, map_location=map_location)
+
+        # users can do some modification after loading checkpoint
+        if callback is not None:
+            callback(self.extra_ckpt)
+
+        state_dict = self.extra_ckpt.pop('state_dict')
+        self.load_model_state_dict(state_dict)
+
+        return self.extra_ckpt
+
+    def resume(
+        self,
+        filename: str,
+        *,
+        resume_optimizer: bool = True,
+        resume_param_scheduler: bool = True,
+        map_location: Union[str, Callable] = 'default',
+        callback: Optional[Callable] = None,
+    ) -> dict:
+        """Resume training from given ``filename``.
+
+        Four types of states will be resumed.
+
+        - model state
+        - optimizer state
+        - scheduler state
+        - randomness state
+
+        Args:
+            filename (str): Accept local filepath, URL, ``torchvision://xxx``,
+                ``open-mmlab://xxx``.
+        """
+        if hasattr(self, 'extra_ckpt'):
+            return self.extra_ckpt
+
+        self.logger.info(f'Resume checkpoint from {filename}')
+
+        self.extra_ckpt = self.load_checkpoint(
+            filename, map_location=map_location, callback=callback)
+
+        if not resume_optimizer:
+            self.extra_ckpt.pop('optimizer', None)
+        else:
+            self.load_optim_state_dict(self.extra_ckpt.pop('optimizer'))
+
+        if not resume_param_scheduler:
+            self.extra_ckpt.pop('param_schedulers', None)
+        else:
+            self.load_scheduler_state_dict(
+                self.extra_ckpt.pop('param_schedulers'))
+
+        # resume random seed
+        resumed_seed = self.extra_ckpt['meta'].get('seed', None)
+        current_seed = self._randomness.get('seed')
+        if resumed_seed is not None and resumed_seed != current_seed:
+            if current_seed is not None:
+                self.logger.warning(f'The value of random seed in the '
+                                    f'checkpoint "{resumed_seed}" is '
+                                    f'different from the value in '
+                                    f'`randomness` config "{current_seed}"')
+            self._randomness.update(seed=resumed_seed)
+            self._set_randomness(**self._randomness)
+
+        # resume iter
+        self.dispatch_kwargs['cur_iter'] = self.extra_ckpt['meta']['iter']
+
+        return self.extra_ckpt
+
+    def save_checkpoint(
+        self,
+        filename: str,
+        *,
+        save_optimizer: bool = True,
+        save_param_scheduler: bool = True,
+        extra_ckpt: Optional[dict] = None,
+        callback: Optional[Callable] = None,
+    ) -> None:
+        """Save checkpoint to given ``filename``.
+
+        Args:
+            filename (str): Filename to save checkpoint.
+            save_optimizer (bool): Whether to save the optimizer to
+                the checkpoint. Defaults to True.
+            save_param_scheduler (bool): Whether to save the param_scheduler
+                to the checkpoint. Defaults to True.
+            extra_ckpt (dict, optional): Extra checkpoint to save.
+                Defaults to None.
+            callback (callable, callable): Callback function to modify the
+                checkpoint before saving the checkpoint.
+                Defaults to None.
+        """
+        from mmengine.runner.checkpoint import save_checkpoint
+
+        state_dict = dict()
+        state_dict['state_dict'] = self.model_state_dict()
+
+        # save optimizer state dict
+        if save_optimizer and hasattr(self, 'optim_wrapper'):
+            state_dict['optimizer'] = self.optim_state_dict()
+
+        # save param scheduler state dict
+        if save_param_scheduler and not hasattr(self, 'param_schedulers'):
+            self.logger.warning(
+                '`save_param_scheduler` is True but strategy has no '
+                'param_schedulers attribute, so skip saving parameter '
+                'schedulers')
+            save_param_scheduler = False
+
+        if save_param_scheduler:
+            state_dict['param_schedulers'] = self.scheduler_state_dict()
+
+        # save extra checkpoint passed by users
+        if extra_ckpt is None:
+            extra_ckpt = dict()
+        if 'meta' not in extra_ckpt:
+            extra_ckpt['meta'] = dict()
+        extra_ckpt['meta'].update(
+            seed=self.seed,
+            time=time.strftime('%Y%m%d_%H%M%S', time.localtime()),
+            mmengine=mmengine.__version__ + get_git_hash(),
+        )
+
+        state_dict.update(extra_ckpt)
+
+        # users can do some modification before saving checkpoint
+        if callback is not None:
+            callback(state_dict)
+
+        save_checkpoint(state_dict, filename)
